Implement a radio friendly mode
The mode is turned on by the option --radio or -u
When in use, analog audio can be fed into the audio part of the
CVBS inputs. No actual video signal is required for the audio
to be encoded. However, a (noisy) dummy video stream is added,
and must be filtered away by a module outside this driver in order
to produce a valid radio service.